cloud_run_code/requirements.txt
numpy
Cython
h5py==2.10.0
Pillow
six
scipy
#opencv-python-headless
opencv-python==4.5.5.64
#opencv-contirb-python
matplotlib
tb-nightly
future
yacs
gdown
flake8
yapf
isort
imageio
tensorflow==1.15
scikit-learn==0.22.2
#keras==2.4.0
keras==2.3.1
torch==1.11.0
torchvision==0.12.0

# The following are added by ashish
#pandas-gbq
#google-cloud-bigquery
google-cloud-bigquery==3.1.0
#google-cloud-bigquery-storage==2.13.1
google-cloud-storage==2.3.0

#Docker
Flask==2.1.0
gunicorn==20.1.0